+++ crates/shell/orbtk_shell_stdweb/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "orbtk_shell_orbclient"
+name = "orbtk_shell_stdweb"
 version = "0.3.1-alpha4"
 authors = ["Florian Blasius <flovanpt@posteo.de>"]
-description = "Window shell crate used by OrbTk. Based on OrbClient"
+description = "Window shell crate used by OrbTk. Based on stdweb."
 repository = "https://github.com/redox-os/orbtk"
 readme = "README.md"
 license = "MIT"
@@ -10,12 +10,8 @@ keywords = ["shell", "window", "ui"]
 edition = "2018"
 
 [dependencies]
-image = { version = "0.23",  default-features = false, features = ["ico"] }
-spin_sleep = { version = "1.0" }
-orbclient = { git = "https://gitlab.redox-os.org/redox-os/orbclient.git" }
+stdweb = "0.4"
 derive_more = { version = "0.99", default-features = false, features = ["constructor"] }
-lazy_static = "1.4.0"
-raw-window-handle = "0.3.3"
 
 # orbtk
 orbtk_shell = { path = "../orbtk_shell", version = "0.3.1-alpha4" }
