+++ crates/shell/orbtk_shell_stdweb/src/lib.rs
@@ -1,0 +1,10 @@
+pub use self::shell::*;
+use self::states::*;
+pub use self::window::*;
+pub use self::window_builder::*;
+
+mod prelude;
+mod shell;
+mod states;
+mod window;
+mod window_builder;
