+++ crates/shell/orbtk_shell_stdweb/src/prelude.rs
@@ -1,0 +1,1 @@
+pub use crate::*;
