+++ crates/shell/orbtk_shell_stdweb/src/window.rs
@@ -10,13 +10,9 @@ use stdweb::{
 use derive_more::Constructor;
 
 use super::EventState;
-use crate::{
-    event::{ButtonState, Key, KeyEvent, MouseButton, MouseEvent},
-    render::RenderContext2D,
-    window_adapter::WindowAdapter,
-    WindowRequest,
-};
 
+use orbtk_render::prelude::*;
+use orbtk_shell::prelude::*;
 use orbtk_utils::Point;
 
 /// Represents a wrapper for a web window. It handles events, propagate them to
